+++ packages/logtape/src/config.ts
@@ -535,6 +535,6 @@ export class ConfigError extends Error {
    */
   constructor(message: string) {
     super(message);
-    this.name = "ConfigureError";
+    this.name = "ConfigError";
   }
 }